thêm data từ plugin ra theme

+++ platform/plugins/product/database/migrations/2024_03_06_145129_product_create_product_items_table.php
@@ -13,7 +13,7 @@ return new class () extends Migration {
             $table->longText('description');
             $table->float('price');
             $table->string('image',555);
-            $table->unsignedBigInteger('category_id',20);
+            $table->integer('category_id',20);
             $table->foreign('category_id')->references('id')->on('product_categories')->onDelete('cascade'); 
             $table->string('status', 60)->default('published');
             $table->timestamps();
